pyproject.toml
[tool.poetry]
name = "b2bbtask"
version = "0.1.0"
description = ""
authors = ["Dmitry O <feanarokurufinve@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-rest-framework = "^0.1.0"
mysqlclient = "^2.2.5"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
black = "^24.10.0"
pytest = "^8.3.3"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}
djangorestframework-stubs = "^3.15.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "b2bbtask.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "b2bbtask.settings"
python_files = ["tests.py"]
